# assets/api-docs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Industry 5.0 RESTful API",
		"description": "\n## Installation\nTo install locally the API, clone the repository and install the dependencies using pip:\n```bash\ngit clone https://github.com/sa-team-d/api.git\ncd api\npip install -r requirements.txt\n```\n\n## Usage\nTo start the API locally, run the following command:\n```bash\nuvicorn main:app --reload\n```\nor use the FastAPI command for development:\n```bash\nfastapi dev main.py\n```\nThe API will be available at `http://localhost:8000/api/<api-version>`.\n\n## Documentation\nThe API documentation is available at `http://localhost:8000/api/<api-version>/docs` as a Swagger UI interface. The documentation provides information about the available endpoints, request and response formats, and examples. For an alternative view, the API documentation is also available at `http://localhost:8000/api/<api-version>/redoc`.\n\n## Plugins\nThe API is organized in plugins, each containing a set of endpoints related to a specific functionality. The available plugins are:\n- `kpi`: Get all KPIs available, filter KPIs by name, and set thresholds\n- `machine`: Get all machines, filter machines by category, and get machine details\n- `user`: Users management\n- `report`: Generate and manage reports\n\n## Contact\nFor any questions or issues, please contact the project maintainers:\n- [Leonardo Stoppani](https://github.com/lilf4p)\n- [Oleksiy Nedobiychuk](https://github.com/lesi-nedo)\n",
		"version": "v1.0"
	},
	"paths": {
		"/": {
			"get": {
				"summary": "Redirect to Swagger docs",
				"operationId": "redirect_to_docs__get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				}
			}
		},
		"/docs": {
			"get": {
				"summary": "Redirect to Swagger docs",
				"operationId": "redirect_to_docs_docs_get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				}
			}
		},
		"/redoc": {
			"get": {
				"summary": "Redirect to ReDoc docs",
				"operationId": "redirect_to_redoc_redoc_get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				}
			}
		},
		"/api/v1.0/machine/": {
			"get": {
				"tags": [
					"Machine"
				],
				"summary": "Get all machines in the dataset",
				"operationId": "get_all_machines_api_v1_0_machine__get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Machine"
									},
									"type": "array",
									"title": "Response Get All Machines Api V1 0 Machine  Get"
								}
							}
						}
					}
				},
				"security": [
					{
						"HTTPBearer": []
					}
				]
			}
		},
		"/api/v1.0/machine/filter": {
			"post": {
				"tags": [
					"Machine"
				],
				"summary": "Filter machines by type or name",
				"operationId": "filter_machines_api_v1_0_machine_filter_post",
				"security": [
					{
						"HTTPBearer": []
					}
				],
				"parameters": [
					{
						"name": "machine_name",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"title": "Machine Name"
						}
					},
					{
						"name": "machine_type",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"title": "Machine Type"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Machine"
									},
									"title": "Response Filter Machines Api V1 0 Machine Filter Post"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1.0/machine/{machine_id}": {
			"get": {
				"tags": [
					"Machine"
				],
				"summary": "Get machine by ID",
				"operationId": "get_machine_by_id_api_v1_0_machine__machine_id__get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Machine"
								}
							}
						}
					}
				},
				"security": [
					{
						"HTTPBearer": []
					}
				]
			}
		},
		"/api/v1.0/user/login": {
			"post": {
				"tags": [
					"User"
				],
				"summary": "TEST ONLY: Simulate frontend login",
				"description": "Authenticate user and get firebase ID token",
				"operationId": "login_api_v1_0_user_login_post",
				"parameters": [
					{
						"name": "email",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Email"
						}
					},
					{
						"name": "password",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Password"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "object",
									"title": "Response Login Api V1 0 User Login Post"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1.0/user/": {
			"post": {
				"tags": [
					"User"
				],
				"summary": "Create a new user",
				"operationId": "create_user_api_v1_0_user__post",
				"parameters": [
					{
						"name": "name",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Name"
						}
					},
					{
						"name": "phone_number",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Phone Number"
						}
					},
					{
						"name": "email",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Email"
						}
					},
					{
						"name": "site",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Site"
						}
					},
					{
						"name": "required_role",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Required Role"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"get": {
				"tags": [
					"User"
				],
				"summary": "Get current user",
				"operationId": "get_current_user_api_v1_0_user__get",
				"security": [
					{
						"HTTPBearer": []
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1.0/user/{user_id}": {
			"delete": {
				"tags": [
					"User"
				],
				"summary": "Delete a user",
				"operationId": "delete_user_api_v1_0_user__user_id__delete",
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "User Id"
						}
					},
					{
						"name": "required_role",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Required Role"
						}
					}
				],
				"responses": {
					"204": {
						"description": "Successful Response"
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"put": {
				"tags": [
					"User"
				],
				"summary": "Update user info",
				"operationId": "update_user_api_v1_0_user__user_id__put",
				"parameters": [
					{
						"name": "user_id",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string",
							"title": "User Id"
						}
					},
					{
						"name": "name",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Name"
						}
					},
					{
						"name": "phone_number",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Phone Number"
						}
					},
					{
						"name": "email",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Email"
						}
					},
					{
						"name": "site",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Site"
						}
					},
					{
						"name": "required_role",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Required Role"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/User"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1.0/user/list": {
			"get": {
				"tags": [
					"User"
				],
				"summary": "List all users",
				"operationId": "list_users_api_v1_0_user_list_get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"type": "array",
									"title": "Response List Users Api V1 0 User List Get"
								}
							}
						}
					}
				},
				"security": [
					{
						"HTTPBearer": []
					}
				]
			}
		},
		"/api/v1.0/user/filter": {
			"post": {
				"tags": [
					"User"
				],
				"summary": "Filter users by name or email",
				"operationId": "filter_users_api_v1_0_user_filter_post",
				"security": [
					{
						"HTTPBearer": []
					}
				],
				"parameters": [
					{
						"name": "first_name",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"title": "First Name"
						}
					},
					{
						"name": "last_name",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"title": "Last Name"
						}
					},
					{
						"name": "email",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"title": "Email"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/User"
									},
									"title": "Response Filter Users Api V1 0 User Filter Post"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1.0/kpi/": {
			"get": {
				"tags": [
					"KPI"
				],
				"summary": "Get all KPIs avaiable",
				"operationId": "get_all_kpis_api_v1_0_kpi__get",
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/KPIGroup"
									},
									"type": "array",
									"title": "Response Get All Kpis Api V1 0 Kpi  Get"
								}
							}
						}
					}
				},
				"security": [
					{
						"HTTPBearer": []
					}
				]
			}
		},
		"/api/v1.0/kpi/filter": {
			"get": {
				"tags": [
					"KPI"
				],
				"summary": "Get all KPIs with a name",
				"operationId": "get_kpis_by_group_api_v1_0_kpi_filter_get",
				"security": [
					{
						"HTTPBearer": []
					}
				],
				"parameters": [
					{
						"name": "kpi_name",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Kpi Name"
						}
					},
					{
						"name": "start_date",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Start Date"
						}
					},
					{
						"name": "end_date",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "End Date"
						}
					},
					{
						"name": "site_name",
						"in": "query",
						"required": false,
						"schema": {
							"type": "string",
							"title": "Site Name"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Kpi"
									},
									"title": "Response Get Kpis By Group Api V1 0 Kpi Filter Get"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1.0/kpi/threshold/": {
			"post": {
				"tags": [
					"KPI"
				],
				"summary": "Set threshold for a KPI",
				"operationId": "set_threshold_api_v1_0_kpi_threshold__post",
				"parameters": [
					{
						"name": "kpi_name",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Kpi Name"
						}
					},
					{
						"name": "threshold",
						"in": "query",
						"required": true,
						"schema": {
							"type": "number",
							"title": "Threshold"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1.0/report/": {
			"post": {
				"tags": [
					"Report"
				],
				"summary": "Create a new report, save it to the database and return it",
				"operationId": "create_report_api_v1_0_report__post",
				"security": [
					{
						"HTTPBearer": []
					}
				],
				"parameters": [
					{
						"name": "name",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Name"
						}
					},
					{
						"name": "site",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Site"
						}
					},
					{
						"name": "kpi",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Kpi"
						}
					},
					{
						"name": "frequency",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Frequency"
						}
					}
				],
				"responses": {
					"201": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Report"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			},
			"get": {
				"tags": [
					"Report"
				],
				"summary": "Get all reports created by the user",
				"operationId": "get_all_reports_api_v1_0_report__get",
				"parameters": [
					{
						"name": "site",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Site"
						}
					},
					{
						"name": "required_role",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Required Role"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Report"
									},
									"title": "Response Get All Reports Api V1 0 Report  Get"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		},
		"/api/v1.0/report/filter": {
			"get": {
				"tags": [
					"Report"
				],
				"summary": "Get all reports for a specific site created by the user",
				"operationId": "get_reports_by_site_api_v1_0_report_filter_get",
				"security": [
					{
						"HTTPBearer": []
					}
				],
				"parameters": [
					{
						"name": "site",
						"in": "query",
						"required": true,
						"schema": {
							"type": "string",
							"title": "Site"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Successful Response",
						"content": {
							"application/json": {
								"schema": {
									"type": "array",
									"items": {
										"$ref": "#/components/schemas/Report"
									},
									"title": "Response Get Reports By Site Api V1 0 Report Filter Get"
								}
							}
						}
					},
					"422": {
						"description": "Validation Error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HTTPValidationError"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"HTTPValidationError": {
				"properties": {
					"detail": {
						"items": {
							"$ref": "#/components/schemas/ValidationError"
						},
						"type": "array",
						"title": "Detail"
					}
				},
				"type": "object",
				"title": "HTTPValidationError"
			},
			"KPIGroup": {
				"properties": {
					"name": {
						"type": "string",
						"title": "Name"
					},
					"kpi_list": {
						"items": {
							"$ref": "#/components/schemas/Kpi"
						},
						"type": "array",
						"title": "Kpi List"
					}
				},
				"type": "object",
				"required": [
					"name",
					"kpi_list"
				],
				"title": "KPIGroup"
			},
			"Kpi": {
				"properties": {
					"kpi_type": {
						"type": "string",
						"title": "Kpi Type"
					},
					"data": {
						"items": {
							"$ref": "#/components/schemas/Value"
						},
						"type": "array",
						"title": "Data"
					},
					"machine_id": {
						"type": "string",
						"title": "Machine Id"
					},
					"config": {
						"type": "object",
						"title": "Config"
					}
				},
				"type": "object",
				"required": [
					"kpi_type",
					"data",
					"machine_id",
					"config"
				],
				"title": "Kpi"
			},
			"Machine": {
				"properties": {
					"id": {
						"type": "string",
						"title": "Id"
					},
					"category": {
						"type": "string",
						"title": "Category"
					},
					"name": {
						"type": "string",
						"title": "Name"
					},
					"kpi_list": {
						"items": {
							"$ref": "#/components/schemas/Kpi"
						},
						"type": "array",
						"title": "Kpi List"
					}
				},
				"type": "object",
				"required": [
					"id",
					"category",
					"name",
					"kpi_list"
				],
				"title": "Machine"
			},
			"Report": {
				"properties": {
					"id": {
						"type": "string",
						"title": "Id"
					},
					"kpi_type": {
						"type": "string",
						"title": "Kpi Type"
					},
					"content": {
						"type": "string",
						"title": "Content"
					},
					"date": {
						"type": "string",
						"format": "date-time",
						"title": "Date"
					},
					"uid": {
						"type": "string",
						"title": "Uid"
					}
				},
				"type": "object",
				"required": [
					"id",
					"kpi_type",
					"content",
					"date",
					"uid"
				],
				"title": "Report"
			},
			"User": {
				"properties": {
					"uid": {
						"type": "string",
						"title": "Uid"
					},
					"email": {
						"type": "string",
						"format": "email",
						"title": "Email"
					},
					"site": {
						"type": "integer",
						"title": "Site"
					},
					"name": {
						"type": "string",
						"title": "Name"
					},
					"phone_number": {
						"type": "string",
						"title": "Phone Number"
					}
				},
				"type": "object",
				"required": [
					"uid",
					"email",
					"site",
					"name",
					"phone_number"
				],
				"title": "User"
			},
			"ValidationError": {
				"properties": {
					"loc": {
						"items": {
							"anyOf": [
								{
									"type": "string"
								},
								{
									"type": "integer"
								}
							]
						},
						"type": "array",
						"title": "Location"
					},
					"msg": {
						"type": "string",
						"title": "Message"
					},
					"type": {
						"type": "string",
						"title": "Error Type"
					}
				},
				"type": "object",
				"required": [
					"loc",
					"msg",
					"type"
				],
				"title": "ValidationError"
			},
			"Value": {
				"properties": {
					"machine_id": {
						"type": "string",
						"title": "Machine Id"
					},
					"sum": {
						"type": "number",
						"title": "Sum"
					},
					"avg": {
						"type": "number",
						"title": "Avg"
					},
					"min": {
						"type": "number",
						"title": "Min"
					},
					"max": {
						"type": "number",
						"title": "Max"
					},
					"var": {
						"type": "number",
						"title": "Var"
					}
				},
				"type": "object",
				"required": [
					"machine_id",
					"sum",
					"avg",
					"min",
					"max",
					"var"
				],
				"title": "Value"
			}
		},
		"securitySchemes": {
			"HTTPBearer": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	}
}